requirements.txt
stripe
python-dotenv
